refactor(auth): migrate to Drizzle adapter with Neon HTTP driver
Better Auth v1.5.x removed the `{ type, url }` database shorthand.
Switch to the officially recommended Drizzle adapter with
@neondatabase/serverless for Cloudflare Workers compatibility.

- Extract shared schema-affecting options into src/auth/options.ts
- Move CLI config from src/auth.ts to better-auth.config.ts (package root)
- Update createAuth() to use drizzleAdapter(drizzle(neon(url)))
- Add @better-auth/drizzle-adapter, drizzle-orm, drizzle-kit deps
- Update CLI scripts to use --config flag with new config path
- Remove tsconfig exclude (CLI config now outside src/)
