+++ packages/server-cloudflare/src/auth/server.ts
@@ -1,50 +1,8 @@
-import type { BetterAuthOptions } from 'better-auth';
+import { neon } from '@neondatabase/serverless';
 import { betterAuth } from 'better-auth';
-import { bearer } from 'better-auth/plugins/bearer';
-import { jwt } from 'better-auth/plugins/jwt';
-import { oauthProvider } from '@better-auth/oauth-provider';
-
-/**
- * Schema-affecting config shared between the runtime auth instance and
- * `src/auth.ts` (CLI migrations). Every option here influences the
- * database schema — keep them in one place so `npx @better-auth/cli migrate`
- * always matches what the worker actually uses.
- *
- * Runtime-only options (secondaryStorage, trustedOrigins, cookies, etc.)
- * belong in `createAuth` below — they don't affect the schema.
- */
-export const sharedAuthConfig = {
-	basePath: '/auth',
-	emailAndPassword: { enabled: true },
-	plugins: [
-		bearer(),
-		jwt(),
-		oauthProvider({
-			loginPage: '/sign-in',
-			consentPage: '/consent',
-			requirePKCE: true,
-			allowDynamicClientRegistration: true,
-			trustedClients: [
-				{
-					clientId: 'epicenter-desktop',
-					name: 'Epicenter Desktop',
-					type: 'native',
-					redirectUrls: ['tauri://localhost/auth/callback'],
-					skipConsent: true,
-					metadata: {},
-				},
-				{
-					clientId: 'epicenter-mobile',
-					name: 'Epicenter Mobile',
-					type: 'native',
-					redirectUrls: ['epicenter://auth/callback'],
-					skipConsent: true,
-					metadata: {},
-				},
-			],
-		}),
-	],
-} satisfies Partial<BetterAuthOptions>;
+import { drizzleAdapter } from 'better-auth/adapters/drizzle';
+import { drizzle } from 'drizzle-orm/neon-http';
+import { authOptions } from './options';
 
 type AuthEnv = {
 	DATABASE_URL: string;
@@ -53,13 +11,23 @@ type AuthEnv = {
 	BETTER_AUTH_URL?: string; // e.g. https://api.epicenter.so — needed for OAuth issuer
 };
 
+/**
+ * Creates a Better Auth instance configured for the Cloudflare Workers runtime.
+ *
+ * Spreads {@link authOptions} (schema-affecting config shared with the CLI) and
+ * layers on runtime-only options: Neon database via Drizzle adapter, Cloudflare
+ * KV session caching, cross-subdomain cookies, and trusted origins.
+ *
+ * Called per-request in `app.ts` — no module-level cache, per Better Auth's
+ * serverless recommendation.
+ */
 export function createAuth(env: AuthEnv) {
+	const sql = neon(env.DATABASE_URL);
+	const db = drizzle(sql);
+
 	const auth = betterAuth({
-		...sharedAuthConfig,
-		database: {
-			type: 'postgres',
-			url: env.DATABASE_URL,
-		},
+		...authOptions,
+		database: drizzleAdapter(db, { provider: 'pg' }),
 		baseURL: env.BETTER_AUTH_URL,
 		secret: env.BETTER_AUTH_SECRET,
 		session: {
